trash_microservices/auth_service/pyproject.toml
[tool.poetry]
name = "auth_service"
version = "0.1.0"
description = ""
authors = ["Arkhip McNutt <ortiz47@mail.ru>"]
readme = "README.md"
include = [
    "auth_service/config/app/app_config.json",
    "auth_service/config/app/jwt_secret.txt",
    "auth_service/config/database/db_config.json",
]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
flask = "^3.0.2"
pypika = "^0.48.9"
pyjwt = "^2.8.0"
pytest = "^8.1.1"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["auth_service"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.json"]

[options]
include_package_data = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
